refactor(components): extract media primitives + shared module (Phase 6 #1)
Splits Image and Icon (with ICON_GLYPHS) out of the 956-line
builtins.tsx into src/components/primitives/media.tsx, and lifts the
cross-family ComponentProps type and resolvedName helper into a new
primitives/shared.ts.

builtins.tsx re-exports Image and Icon so existing consumers
(A2UIRenderer, builtins.test.tsx) keep working with no source changes.
No logic changes — pure file organization. Refs #43.
